LPS-55581 Moves EditUserGroupAssignmentsAction - no logic changes
